% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_performance.R
\name{brier_score,Landmarking-method}
\alias{brier_score,Landmarking-method}
\title{Computes Brier score at the desired landmark times and prediction windows.}
\usage{
\S4method{brier_score}{Landmarking}(x, landmarks, windows)
}
\arguments{
\item{x}{An object of class \code{\link{Landmarking}}.}

\item{landmarks}{A numeric vector of landmark times.}

\item{windows}{A vector of prediction windows determining horizon times.}
}
\description{
Computes Brier score at the desired landmark times and prediction windows.
}
