% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_performance.R
\name{brier_score}
\alias{brier_score}
\title{Brier Score}
\usage{
brier_score(x, landmarks, windows)
}
\arguments{
\item{windows}{}
}
\description{
Computes concordance index (c-index) at the desired landmark times and
prediction windows.
}
