% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_performance.R
\name{c_index,Landmarking-method}
\alias{c_index,Landmarking-method}
\title{Computes concordance index (c-index) at the desired landmark times and
prediction windows.}
\usage{
\S4method{c_index}{Landmarking}(x, landmarks, windows)
}
\arguments{
\item{x}{An object of class \code{\link{Landmarking}}.}

\item{landmarks}{A numeric vector of landmark times.}

\item{windows}{A vector of prediction windows determining horizon times.}
}
\description{
Computes concordance index (c-index) at the desired landmark times and
prediction windows.
}
