% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_performance.R
\name{c_index}
\alias{c_index}
\title{C-index}
\usage{
c_index(x, landmarks, windows)
}
\arguments{
\item{windows}{}
}
\description{
Computes concordance index (c-index) at the desired landmark times and
prediction windows.
}
