% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_performance_aux.R
\name{CIndexCRisks}
\alias{CIndexCRisks}
\title{Concordance index for competing risks}
\usage{
CIndexCRisks(
  predictions,
  time,
  cens.code = 0,
  status,
  cause,
  tau,
  method = c("survival", "cifs")
)

CIndexCRisks(
  predictions,
  time,
  cens.code,
  status,
  cause,
  tau,
  method = c("survival", "cifs")
)
}
\arguments{
\item{predictions}{Numeric vector of model predictions.}

\item{time}{Numeric vector describing the time to the event of interest or censoring.}

\item{cens.code}{Value used to denote censoring in \code{status}. Defaults to
0.}

\item{status}{Vector of censoring status.}

\item{cause}{Event of interest.}

\item{tau}{Time c-index is evaluated.}

\item{method}{\code{'survival'} if the predictions are survival probabilities
or \code{'cifs'} if they are cumulative incidence functions}
}
\value{
Concordance index value.
}
\description{
Assess discriminative performance of predictions obtained from a conventional
or competing risks time-to-event model using time-dependent concordance index.
}
\details{
Uses the proportion of correctly ordered risk pairs for the event \eqn{k},
based on the predicted risk of the event up to time \eqn{\tau}.

\deqn{{C_k(\tau) = \frac{\sum_{i=1}^N \sum_{j=1}^N (A_{ij} + B_{ij}) \cdot Q_{ij} \cdot N_i^k(\tau)}{\sum_{i=1}^N \sum_{j=1}^N (A_{ij} + B_{ij}) \cdot N_i^k(\tau)}}}

A == risk ordering of patients, small time means patient 'i' at higher risk
than patient 'j' experiencing event of interest \eqn{A[i,j] = 0} for tied
event times.

B == risk ordering of patients, large time for patient 'i' means lower risk
than patient 'j' if not experienced the event of interest. Ties are included
in B

Q == the risk ordering of the subjects, i.e., is subject i assigned a higher
risk by the model than the subject j, for event \eqn{E_k} until time \eqn{t}.
\eqn{Q[i,j] = 0} for tied predictions.

N_t == number of subjects with survival time < time point and experience event of interest
Tied event times are included
}
\references{
Ahuja K, Schaar M van der. Joint Concordance Index. Published online August 17, 2019. doi:\href{https://doi.org/10.48550/arXiv.1810.11207}{10.48550/arXiv.1810.11207}
}
