% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\name{compute_risk_sets,Landmarking-method}
\alias{compute_risk_sets,Landmarking-method}
\title{Compute the list of individuals at risk at given landmark times,
and stores them in an object of class Landmarking}
\usage{
\S4method{compute_risk_sets}{Landmarking}(x, landmarks, ...)
}
\arguments{
\item{x}{An object of class \code{\link{Landmarking}}.}

\item{landmarks}{Numeric vector of landmark times}

\item{...}{Additional arguments (not used)}
}
\value{
An object of class Landmarking, including desired risk sets for the
relevant landmark times.
}
\description{
Compute the list of individuals at risk at given landmark times,
and stores them in an object of class Landmarking
}
