% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\name{compute_risk_sets,LandmarkingCV-method}
\alias{compute_risk_sets,LandmarkingCV-method}
\title{Compute the list of individuals at risk at given landmark times,
and stores them in an object of class LandmarkingCV}
\usage{
\S4method{compute_risk_sets}{LandmarkingCV}(x, landmarks, K = 1)
}
\arguments{
\item{x}{An object of class \code{\link{Landmarking}}.}

\item{landmarks}{Numeric vector of landmark times}

\item{K}{Number of cross-validation folds.}
}
\value{
An object of class LandmarkingCV, including desired risk sets for the
relevant landmark times.
}
\description{
Compute the list of individuals at risk at given landmark times,
and stores them in an object of class LandmarkingCV
}
