% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_lcmm.R
\name{fit_lcmm_}
\alias{fit_lcmm_}
\title{Fits an LCMM model}
\usage{
fit_lcmm_(formula, data, mixture, subject, ng, ...)
}
\arguments{
\item{formula}{Two-sided linear formula for the fixed effects in the LCMM.}

\item{data}{Dataframe with data}

\item{mixture}{One-sided formula specifying the class-specific fixed effects.}

\item{subject}{Name of the column indicating individual ids in data}

\item{ng}{Number of clusters in the LCMM model}

\item{...}{Additional arguments passed to the lcmm::hlme function.}
}
\value{
An object of class hlme
}
\description{
Fits an LCMM model
}
