% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_longitudinal.R
\name{fit_longitudinal_v2}
\alias{fit_longitudinal_v2}
\title{Fits the specified longitudinal model for the latent processes underlying the
relevant time-varying covariates, up until the landmarking times}
\usage{
fit_longitudinal_v2(x, landmarks, method, formula, ...)
}
\arguments{
\item{x}{An object of class Landmarking.}

\item{landmarks}{A vector of Landmark times.}

\item{method}{A function for fitting a longitudinal data model, whose first
argument is a formula, and has at least an argument "data".}

\item{formula}{A formula to be used in longitudinal submodel fitting.}

\item{...}{Additional arguments passed to the longitudinal model fitting function (e.g. number of classes/clusters for lcmm).}
}
\value{
An object of class Landmarking.
}
\description{
Fits the specified longitudinal model for the latent processes underlying the
relevant time-varying covariates, up until the landmarking times
}
