% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_survival.R
\name{fit_survival}
\alias{fit_survival}
\title{Fits the specified survival model at the landmark times and up to the horizon
times specified by the user}
\usage{
fit_survival(x, landmarks, windows, method, dynamic_covariates = c())
}
\arguments{
\item{x}{An object of class \code{\link{Landmarking}}.}

\item{landmarks}{Numeric vector of landmark times}

\item{windows}{Vector of prediction windows determining horizon times.}

\item{method}{Method for survival analysis, either "survfit" or "coxph".}

\item{dynamic_covariates}{Vector of time-varying covariates to be used
in the survival model.}
}
\value{
An object of class Landmarking.
}
\description{
Fits the specified survival model at the landmark times and up to the horizon
times specified by the user
}
