% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\docType{class}
\name{Landmarking-class}
\alias{Landmarking-class}
\title{An S4 class for a landmarking analysis}
\description{
An S4 class for a landmarking analysis
}
\section{Slots}{

\describe{
\item{\code{landmarks}}{Array of landmark times.}

\item{\code{data_static}}{Data frame with id, static covariates, censoring indicator and event time.}

\item{\code{data_dynamic}}{Data frame in long format with id, measurement, measurement time and covariate measured.}

\item{\code{event_indicator}}{Name of the column indicating event or censoring.}

\item{\code{biomarkers}}{List whose name indicate columns for biomarker measurements, and values indicate columns for biomarker measurement times}

\item{\code{ids}}{Name of the column indicating patient id.}

\item{\code{event_time}}{Name of the column indicating time of the event/censoring.}

\item{\code{risk_sets}}{List of indices.}

\item{\code{longitudinal_fits}}{List of model fits for the specified landmark times and biomarkers.}

\item{\code{longitudinal_predictions}}{List of model predictions for the specified landmark times and biomarkers.}

\item{\code{survival_fits}}{List of survival model fits at each of the specified landmark times.}
}}

