% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking-cv.R
\docType{class}
\name{LandmarkingCV-class}
\alias{LandmarkingCV-class}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{folds}}{Vector of integers associating observations to CV folds.}

\item{\code{landmarking_list}}{List of Landmarking objects, one per CV fold.}
}}

