% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarkR-package.R
\docType{package}
\name{landmarkR}
\alias{landmarkR-package}
\alias{landmarkR}
\title{landmarkR}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a comprehensive framework for conducting time-to-event landmarking analysis using a variety of longitudinal and survival sub-models. It aims to facilitate the analysis of time-to-event data with time-dependent covariates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vallejosgroup.github.io/landmarkR/}
  \item \url{https://github.com/VallejosGroup/landmarkR}
  \item Report bugs at \url{https://github.com/VallejosGroup/landmarkR/issues}
}

}
\author{
\strong{Maintainer}: Nathan Constantine-Cooke \email{nathan.constantine-cooke@ed.ac.uk} (\href{https://orcid.org/0000-0002-4437-8713}{ORCID})

Authors:
\itemize{
  \item Victor Velasco-Pardo \email{vvelasco@ed.ac.uk} (\href{https://orcid.org/0000-0002-7166-1573}{ORCID})
  \item Catalina Vallejos \email{catalina.vallejos@ed.ac.uk} (\href{https://orcid.org/0000-0003-3638-1960}{ORCID})
  \item Charlie Lees \email{charlie.lees@ed.ac.uk} (\href{https://orcid.org/0000-0002-0732-8215}{ORCID})
}

Other contributors:
\itemize{
  \item Begoña Bolos Sierra [contributor]
}

}
\keyword{internal}
