% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_performance.R
\name{performance_metrics}
\alias{performance_metrics}
\title{Performance metrics}
\usage{
performance_metrics(x, landmarks, windows, c_index = TRUE, brier = TRUE)
}
\arguments{
\item{x}{Object of class Landmarking.}

\item{landmarks}{Numeric vector of landmark times.}

\item{windows}{Vector of prediction windows determining horizon times.}

\item{c_index}{Logical. If TRUE (default), C index is reported.}

\item{brier}{Logical. If TRUE (default), Brier score is reported.}
}
\value{
Data frame with performance metrics across the specified landmark
times and prediction windows.
}
\description{
Computes concordance index (c-index) and Brier scores at the specified landmark
times and prediction windows.
}
