% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_lcmm.R
\name{predict_lcmm_}
\alias{predict_lcmm_}
\title{Makes predictions from an LCMM model}
\usage{
predict_lcmm_(x, newdata, subject, avg = FALSE)
}
\arguments{
\item{x}{An object of class \code{\link{hlme}{lcmm}}.}

\item{newdata}{A data frame containing static covariates and individual
IDs}

\item{subject}{Name of the column in newdata where individual IDs are stored.}

\item{avg}{Boolean indicating whether to make predictions based on the
most likely cluster (FALSE, default) or averaging over clusters (TRUE).}
}
\value{
A vector of predictions.
}
\description{
Makes predictions from an LCMM model
}
