% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_longitudinal.R
\name{predict_longitudinal,Landmarking-method}
\alias{predict_longitudinal,Landmarking-method}
\title{Make predictions for time-varying covariates at specified landmark times}
\usage{
\S4method{predict_longitudinal}{Landmarking}(x, landmarks, method, static_covariates, ...)
}
\arguments{
\item{x}{An object of class Landmarking.}

\item{landmarks}{Vector of landmark times.}

\item{method}{Longitudinal data analysis method used to make predictions}

\item{static_covariates}{Vector of names of static covariates to be included in the longitudinal analysis.}

\item{...}{Additional arguments passed to the prediction function (e.g. number of classes/clusters for lcmm).}
}
\value{
An object of class Landmarking
}
\description{
Make predictions for time-varying covariates at specified landmark times
}
