% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_longitudinal.R
\name{predict_longitudinal}
\alias{predict_longitudinal}
\title{Make predictions for time-varying covariates at specified landmark times}
\usage{
predict_longitudinal(x, landmarks, method, ...)
}
\arguments{
\item{x}{An object of class Landmarking.}

\item{landmarks}{Vector of landmark times.}

\item{method}{Longitudinal data analysis method used to make predictions}

\item{...}{Additional arguments passed to the prediction function (e.g. number of classes/clusters for lcmm).}
}
\value{
An object of class Landmarking
}
\description{
Make predictions for time-varying covariates at specified landmark times
}
