% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_split_wide_df.R
\name{split_wide_df}
\alias{split_wide_df}
\title{Separate a wide format data frame describing static and dynamic covariates
into long format data frames}
\usage{
split_wide_df(df, ids, times, static, dynamic, measurement_name)
}
\arguments{
\item{df}{A data frame in wide format.}

\item{ids}{The name of the column that identifies individuals in \code{df}.}

\item{times}{The name of the column that identifies measurement times
in \code{df}.}

\item{static}{A vector with the column names in \code{df} that store static
covariates.}

\item{dynamic}{A vector with the column names in \code{df} that store dynamic
covariates.}

\item{measurement_name}{The name for the columns where values of
dynamic covariates will be stored.}
}
\value{
A list containing df_static, a data frame of all static observations
and  df_dynamic, a list of data frames, one per dynamic covariate.
}
\description{
Splits a wide format data frame into a data frame of static covariates and a
list of data frames where each element of the list is associated with a
dynamic covariate.
}
\examples{
data(epileptic)
epileptic_dfs <- split_wide_df(
  epileptic,
  ids = "id",
  times = "time",
  static = c("with.time", "with.status", "treat", "age", "gender", "learn.dis"),
  dynamic = c("dose"),
  measurement_name = "value"
)
}
