% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_performance_aux.R
\name{BinaryBrierScore}
\alias{BinaryBrierScore}
\title{Binary Brier Score}
\usage{
BinaryBrierScore(predictions, time, status, tau, cause)
}
\arguments{
\item{predictions}{Numeric vector of model predictions.}

\item{time}{time vector of times. Time-to-event.}

\item{status}{vector of events. 1) If survival -> \link{0,...,1}. 2) If competing risks -> \link{0,...,K}.}

\item{tau}{Time Brier score is evaluated.}

\item{cause}{event of interest.}
}
\description{
Computes the Binary Brier Score (BBS) for binary outcomes at a specified
time point \eqn{\tau}.
}
\details{
The BBS is defined as the mean squared difference between the predicted
probabilities and the true outcome.
}
