% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_performance.R
\name{c_index}
\alias{c_index}
\title{C-index}
\usage{
c_index(x, landmarks, windows)
}
\arguments{
\item{x}{An object of class \code{\link{Landmarking}}.}

\item{landmarks}{A numeric vector of landmark times.}

\item{windows}{A vector of prediction windows determining horizon times.}
}
\description{
Computes concordance index (c-index) at the desired landmark times and
prediction windows.
}
