% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\name{compute_risk_sets}
\alias{compute_risk_sets}
\title{Compute the list of individuals at risk at given landmark times,
and stores them in an object of class Landmarking}
\usage{
compute_risk_sets(x, landmarks, K = 1)
}
\arguments{
\item{x}{An object of class \code{\link{Landmarking}}.}

\item{landmarks}{Numeric vector of landmark times}

\item{K}{Number of cross-validation folds.}

\item{...}{Additional arguments (not used)}
}
\value{
An object of class \code{\link{Landmarking}} including desired risk
sets for the specified landmark times.
}
\description{
Compute the list of individuals at risk at given landmark times,
and stores them in an object of class Landmarking
}
