% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\name{compute_risk_sets}
\alias{compute_risk_sets}
\title{Compute the list of individuals at risk at given landmark times,
and stores them in an object of class Landmarking}
\usage{
compute_risk_sets(x, landmarks, ...)
}
\arguments{
\item{x}{An object of type Landmarking}

\item{landmarks}{Vector of landmark times}

\item{...}{Additional arguments (not used)}
}
\value{
An
}
\description{
Compute the list of individuals at risk at given landmark times,
and stores them in an object of class Landmarking
}
