% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_longitudinal.R
\name{fit_longitudinal,Landmarking-method}
\alias{fit_longitudinal,Landmarking-method}
\title{Fit specified longitudinal models for time-dependent covariates}
\usage{
\S4method{fit_longitudinal}{Landmarking}(x, landmarks, method, formula, ...)
}
\arguments{
\item{x}{An object of class \code{\link{Landmarking}}.}

\item{landmarks}{A vector of Landmark times.}

\item{method}{A function for fitting a longitudinal data model, whose first
argument is a formula, and has at least an argument "data".}

\item{formula}{A formula to be used in longitudinal submodel fitting.}

\item{...}{Additional arguments passed to the longitudinal model fitting
function (e.g. number of classes/clusters for lcmm).}
}
\value{
An object of class \code{\link{Landmarking}}.
}
\description{
Fits the specified longitudinal model for the latent processes underlying the
relevant time-varying covariates up to the landmark time(s)
}
\examples{
# Fit a linear mixed-effects model

data(epileptic)

epileptic_dfs <- split_wide_df(
  epileptic,
  ids = "id",
  times = "time",
  static = c("with.time", "with.status", "treat", "age", "gender", "learn.dis"),
  dynamic = c("dose"),
  measurement_name = "value"
)

landmarking_object <- Landmarking(
  data_static = epileptic_dfs$df_static,
  data_dynamic = epileptic_dfs$df_dynamic,
  event_indicator = "with.status",
  ids = "id",
  event_time = "with.time",
  times = "time",
  measurements = "value"
  ) |>
  compute_risk_sets(landmarks = seq(from = 365.25, to = 5 * 365.25, by = 365.25)) |>
  fit_longitudinal(
    landmarks = seq(from = 365.25, to = 5 * 365.25, by = 365.25),
    method = "lme4",
    formula = value ~ treat + age + gender + learn.dis + (time | id)
  )
)
}
