% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_longitudinal.R
\name{fit_longitudinal,Landmarking-method}
\alias{fit_longitudinal,Landmarking-method}
\title{Fits the specified longitudinal model for the latent processes underlying the
relevant time-varying covariates, up until the landmarking times}
\usage{
\S4method{fit_longitudinal}{Landmarking}(x, landmarks, method, static_covariates, ...)
}
\arguments{
\item{x}{An object of class Landmarking.}

\item{landmarks}{A vector of Landmark times.}

\item{method}{Method for longitudinal analysis, currently only "lme4" or "lcmm" is supported.}

\item{static_covariates}{Vector of names of static covariates to be included in the longitudinal analysis.}

\item{...}{Additional arguments passed to the longitudinal model fitting function (e.g. number of classes/clusters for lcmm).}
}
\value{
An object of class Landmarking.
}
\description{
Fits the specified longitudinal model for the latent processes underlying the
relevant time-varying covariates, up until the landmarking times
}
