% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\name{fit_longitudinal}
\alias{fit_longitudinal}
\title{Title}
\usage{
fit_longitudinal(x, landmarks, method, static_covariates)
}
\arguments{
\item{landmarks}{Vector of landmark times.}

\item{method}{Method for longitudinal analysis, currently only "lme4" is supported.}

\item{static_covariates}{Vector of names of static covariates to be included in the longitudinal analysis.}
}
\description{
Title
}
