% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarkeR-package.R
\docType{package}
\name{landmarkeR}
\alias{landmarkeR-package}
\alias{landmarkeR}
\title{landmarkeR}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Nathan Constantine-Cooke \email{nathan.constantine-cooke@ed.ac.uk}

Authors:
\itemize{
  \item Victor Velasco-Pardo \email{vvelasco@ed.ac.uk}
  \item Catalina Vallejos \email{catalina.vallejos@ed.ac.uk}
  \item Charlie Lees \email{charlie.lees@ed.ac.uk}
}

}
\keyword{internal}
