% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\docType{class}
\name{Landmarking-class}
\alias{Landmarking-class}
\title{S4 class for performing a landmarking analysis}
\description{
S4 class for performing a landmarking analysis
}
\section{Slots}{

\describe{
\item{\code{landmarks}}{A numeric vector of landmark times.}

\item{\code{data_static}}{A data frame containing subject ids, static covariates,}

\item{\code{data_dynamic}}{Data frame in long format with subject ids, measurement
values, measurement times and measurement name.}

\item{\code{event_indicator}}{Name of the column indicating event or censoring.}

\item{\code{ids}}{Name of the column indicating subject ids.}

\item{\code{times}}{Name of the column indicating observation time in
\code{data_dynamic}.}

\item{\code{measurements}}{Name of the column indicating measurement values in
\code{data_dynamic}.}

\item{\code{event_time}}{Name of the column indicating time of the event/censoring.}

\item{\code{risk_sets}}{List of indices.}

\item{\code{longitudinal_fits}}{List of model fits for the specified landmark times
and biomarkers.}

\item{\code{longitudinal_predictions}}{List of model predictions for the specified
landmark times and biomarkers.}

\item{\code{survival_fits}}{List of survival model fits at each of the specified
landmark times.}

\item{\code{survival_predictions}}{List of time-to-eventpredictions for the specified
landmark times and prediction windows.}
}}

