% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\name{Landmarking}
\alias{Landmarking}
\title{Creates an S4 class for a landmarking analysis}
\usage{
Landmarking(
  data_static,
  data_dynamic,
  event_indicator,
  ids,
  event_time,
  times,
  measurements
)
}
\arguments{
\item{data_static}{A data frame containing subject ids, static covariates,}

\item{data_dynamic}{Data frame in long format with subject ids, measurement
values, measurement times and measurement name.}

\item{event_indicator}{Name of the column indicating event or censoring.}

\item{ids}{Name of the column indicating subject ids.}

\item{event_time}{Name of the column indicating time of the event/censoring.}

\item{times}{Name of the column indicating observation time in
\code{data_dynamic}.}

\item{measurements}{Name of the column indicating measurement values in
\code{data_dynamic}.}
}
\value{
An object of class Landmarking
}
\description{
Creates an S4 class for a landmarking analysis
}
\examples{
data(epileptic)

epileptic_dfs <- split_wide_df(
  epileptic,
  ids = "id",
  times = "time",
  static = c("with.time", "with.status", "treat", "age", "gender", "learn.dis"),
  dynamic = c("dose"),
  measurement_name = "value"
)

landmarking_object <- Landmarking(
  data_static = epileptic_dfs$df_static,
  data_dynamic = epileptic_dfs$df_dynamic,
  event_indicator = "with.status",
  ids = "id",
  event_time = "with.time",
  times = "time",
  measurements = "value"
)


}
