% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\name{Landmarking}
\alias{Landmarking}
\title{Creates an S4 class for a landmarking analysis}
\usage{
Landmarking(
  data_static,
  data_dynamic,
  event_indicator,
  dynamic_covariates,
  ids,
  event_time,
  times,
  measurements,
  dynamic_covariate_names
)
}
\arguments{
\item{data_static}{Data frame with id, static covariates, censoring indicator and event time.}

\item{data_dynamic}{Data frame in long format with id, measurement, measurement time and covariate measured.}

\item{event_indicator}{Name of the column indicating event or censoring.}

\item{dynamic_covariates}{Names of dynamic (time-varying) covariates.}

\item{ids}{Name of the column indicating patient id.}

\item{event_time}{Name of the column indicating time of the event/censoring.}

\item{times}{Name of the column indicating time in the dynamic df.}

\item{measurements}{Name of the column indicating measurement values in the dynamic df.}

\item{dynamic_covariate_names}{Name of the column indicating names of the dynamic covariates in the dynamic df.}
}
\value{
An object of class Landmarking
}
\description{
Creates an S4 class for a landmarking analysis
}
