% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarkR-package.R
\docType{package}
\name{landmarkR}
\alias{landmarkR-package}
\alias{landmarkR}
\title{landmarkR}
\description{
Provides a comprehensive framework for conducting time-to-event landmarking analysis using a variety of longitudinal and survival sub-models. It aims to facilitate the analysis of time-to-event data with time-dependent covariates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/VallejosGroup/landmarkR}
  \item Report bugs at \url{https://github.com/VallejosGroup/landmarkR/issues}
}

}
\author{
\strong{Maintainer}: Nathan Constantine-Cooke \email{nathan.constantine-cooke@ed.ac.uk}

Authors:
\itemize{
  \item Victor Velasco-Pardo \email{vvelasco@ed.ac.uk}
  \item Catalina Vallejos \email{catalina.vallejos@ed.ac.uk}
  \item Charlie Lees \email{charlie.lees@ed.ac.uk}
}

}
\keyword{internal}
