% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_longitudinal.R
\name{predict_longitudinal,Landmarking-method}
\alias{predict_longitudinal,Landmarking-method}
\title{Make predictions for time-varying covariates at specified landmark times}
\usage{
\S4method{predict_longitudinal}{Landmarking}(x, landmarks, method, ...)
}
\arguments{
\item{x}{An object of class Landmarking.}

\item{landmarks}{Vector of landmark times.}

\item{method}{Function used to make predictions, whose first argument is a
model fit, and with at least an additional argument newdata.}

\item{...}{Additional arguments passed to the prediction function (e.g. number of classes/clusters for lcmm).}
}
\value{
An object of class Landmarking
}
\description{
Make predictions for time-varying covariates at specified landmark times
}
