% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking_split_wide_df.R
\name{split_wide_df}
\alias{split_wide_df}
\title{Split a wide dataframe containing static and dynamic covariates and splits
in into a dataframe with the static covariates and a list of dataframes,
each associated to a dynamic covariate.}
\usage{
split_wide_df(df, ids, times, static, dynamic)
}
\arguments{
\item{df}{A dataframe in wide format.}

\item{ids}{The name of the column that identifies individuals in \code{df}.}

\item{times}{The name of the column that identifies measurement times
in \code{df}.}

\item{static}{A vector with the column names in df that store static
covariates.}

\item{dynamic}{A vector with the column names in df that store dynamic
covariates.}
}
\value{
A dataframe with the static covariates, and a list of dataframes,
one per dynamic covariate.
}
\description{
Split a wide dataframe containing static and dynamic covariates and splits
in into a dataframe with the static covariates and a list of dataframes,
each associated to a dynamic covariate.
}
